% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly.kernel.R
\name{poly.kernel}
\alias{poly.kernel}
\title{Compute the kernel matrix for ROC-SVM path}
\usage{
poly.kernel(x, y = x, param.kernel = 1, ...)
}
\arguments{
\item{x}{An n x p matrix of features}

\item{y}{An m x p matrix of features}

\item{param.kernel}{The parameter(s) for the kernel. For the radial kernel, the parameter is known in the fields as "gamma". For the polynomial kernel, it is the "degree"}

\item{...}{unused}
}
\description{
Compute the kernel matrix for ROC-SVM path. This function comes from \code{svmpath} package by Trevor Hastie. If you want to know details of this function, refer the \code{svmpath} package.
}
